Refactor zoom bahvior into own file
