+++ react/src/components/Map/Map.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect, useMemo, useCallback } from 'react';
-import { useLocation } from 'react-router-dom';
+import React, { useMemo } from 'react';
 import {
   MapContainer,
   ZoomControl,
@@ -7,29 +6,18 @@ import {
   Popup,
   TileLayer,
   WMSTileLayer,
-  useMap,
 } from 'react-leaflet';
 import { TiledMapLayer } from 'react-esri-leaflet';
 import MarkerClusterGroup from '@changey/react-leaflet-markercluster';
-import { TileServerLayer, FeatureCollection, Feature } from '@hazmapper/types';
 import * as L from 'leaflet';
-import * as turf from '@turf/turf';
-import { LatLngTuple, MarkerCluster } from 'leaflet';
+import { MarkerCluster } from 'leaflet';
 import 'leaflet.markercluster';
 
 import 'leaflet/dist/leaflet.css';
 
-const MAP_CONFIG = {
-  startingCenter: [40, -80] as LatLngTuple,
-  maxFitBoundsInitialZoom: 18,
-  maxFitBoundsSelectedFeatureZoom: 18,
-  maxBounds: [
-    [-90, -180], // Southwest coordinates
-    [90, 180], // Northeast coordinates
-  ] as L.LatLngBoundsExpression,
-  minZoom: 2, // 2 typically prevents zooming out too far to see multiple earths
-  maxZoom: 24,
-} as const;
+import { TileServerLayer, FeatureCollection } from '@hazmapper/types';
+import { MAP_CONFIG } from './config';
+import FitBoundsHandler from './FitBoundsHandler';
 
 interface LeafletMapProps {
   /**
@@ -48,64 +36,6 @@ const ClusterMarkerIcon = (childCount: number) => {
     html: `<div><b>${childCount}</b></div>`,
     className: 'marker-cluster',
   });
-};
-
-/**
- * Handles map bounds adjustments based on features.
- * On initial load: Fits bounds to show all features in collection
- * When activeFeatureId changes: Zooms to that specific feature
- * Uses turf.bbox() for bounding calculations with maxFitToBoundsZoom limit
- */
-const FitBoundsHandler: React.FC<{ featureCollection: FeatureCollection }> = ({
-  featureCollection,
-}) => {
-  const location = useLocation();
-  const map = useMap();
-
-  const selectedFeatureId = useMemo(() => {
-    const searchParams = new URLSearchParams(location.search);
-    const rawId = searchParams.get('selectedFeature');
-    return rawId ? Number(rawId) : undefined;
-  }, [location.search]);
-
-  const getBoundsFromFeature = useCallback(
-    (feature: FeatureCollection | Feature) => {
-      const bbox = turf.bbox(feature);
-      return [
-        [bbox[1], bbox[0]] as [number, number],
-        [bbox[3], bbox[2]] as [number, number],
-      ];
-    },
-    []
-  );
-
-  // Handle initial bounds
-  useEffect(() => {
-    if (featureCollection.features.length && !selectedFeatureId) {
-      const bounds = getBoundsFromFeature(featureCollection);
-      map.fitBounds(bounds, {
-        maxZoom: MAP_CONFIG.maxFitBoundsInitialZoom,
-      });
-    }
-  }, [map, featureCollection, selectedFeatureId]);
-
-  // Handle selected feature bounds
-  useEffect(() => {
-    if (selectedFeatureId) {
-      const activeFeature = featureCollection.features.find(
-        (f) => f.id === selectedFeatureId
-      );
-
-      if (activeFeature) {
-        const bounds = getBoundsFromFeature(activeFeature);
-        map.fitBounds(bounds, {
-          maxZoom: MAP_CONFIG.maxFitBoundsSelectedFeatureZoom,
-        });
-      }
-    }
-  }, [map, selectedFeatureId, featureCollection]);
-
-  return null;
 };
 
 /**
